Implement shutdown on server
This is nastier than it looks like: We put an atom with the system into
a component in the system, so that we can stop ourselves. To make this
work with reloaded.repl, we just reimplement it but with atoms instead.
This means we can perform shutdowns on a dev environment and can still
start it again with (start).