Start inlein server if not already running
This has its disadvantages, and it should be refined somewhat in later
versions. Right now, there's a race condition if inlein is executed
concurrently. We should add in a lockfile or something to message other
inlein clients that a server is spinning up, so that they too won't spin
up inlein daemons.